+++ playground/MVFC.RazorRender.Playground.Api/Components/InvoiceEmail.razor.cs
@@ -1,0 +1,7 @@
+﻿namespace MVFC.RazorRender.Playground.Api.Components;
+
+public partial class InvoiceEmail : ComponentBase
+{
+    [Parameter]
+    public required InvoiceModel Model { get; set; }
+}