+++ playground/MVFC.RazorRender.Playground.Api/Components/WelcomeEmail.razor.cs
@@ -1,0 +1,7 @@
+﻿namespace MVFC.RazorRender.Playground.Api.Components;
+
+public partial class WelcomeEmail : ComponentBase
+{
+    [Parameter]
+    public required WelcomeEmailModel Model { get; set; }
+}