+++ playground/MVFC.RazorRender.Playground.Api/Parameters/InvoiceParameters.cs
@@ -1,0 +1,8 @@
+﻿namespace MVFC.RazorRender.Playground.Api.Parameters;
+
+public sealed class InvoiceParameters : IRazorCacheParameter
+{
+    public InvoiceModel Model { get; set; } = default!;
+
+    public string CacheKey => $"invoice-{Model.InvoiceNumber}";
+}