+++ playground/MVFC.RazorRender.Playground.Api/Parameters/WelcomeEmailParameters.cs
@@ -1,0 +1,8 @@
+﻿namespace MVFC.RazorRender.Playground.Api.Parameters;
+
+public sealed class WelcomeEmailParameters : IRazorCacheParameter
+{
+    public WelcomeEmailModel Model { get; set; } = default!;
+
+    public string CacheKey => $"welcome-{Model.Email}";
+}