+++ playground/MVFC.RazorRender.Playground.Api/Usings.cs
@@ -1,0 +1,9 @@
+﻿global using MVFC.RazorRender.Extensions;
+global using MVFC.RazorRender.Interfaces.Parameters;
+global using MVFC.RazorRender.Playground.Api.Models;
+global using MVFC.RazorRender.Interfaces.Services;
+global using MVFC.RazorRender.Playground.Api.Components;
+global using MVFC.RazorRender.Playground.Api.Parameters;
+global using MVFC.RazorRender.Playground.Api.Endpoints;
+global using Microsoft.AspNetCore.Components;
+global using MVFC.RazorRender.Playground.Api.Mock;