+++ tests/MVFC.RazorRender.Tests/Fixture/ProjectAppHost.cs
@@ -1,0 +1,19 @@
+﻿namespace MVFC.RazorRender.Tests.Fixture;
+
+internal sealed class ProjectAppHost() : DistributedApplicationFactory(typeof(MVFC_RazorRender_Playground_Api))
+{
+    protected override void OnBuilderCreating(DistributedApplicationOptions applicationOptions, HostApplicationBuilderSettings hostOptions)
+    {
+        applicationOptions.AllowUnsecuredTransport = true;
+        hostOptions.Args = ["--testmode=true"];
+    }
+
+    protected override void OnBuilderCreated(DistributedApplicationBuilder builder) =>
+        builder.Services.ConfigureHttpClientDefaults(x => x.AddStandardResilienceHandler(c =>
+        {
+            var timeSpan = TimeSpan.FromMinutes(2);
+            c.AttemptTimeout.Timeout = timeSpan;
+            c.CircuitBreaker.SamplingDuration = timeSpan * 2;
+            c.TotalRequestTimeout.Timeout = timeSpan * 3;
+        }));
+}